requirements.txt
requests>=2.31.0
edge-tts>=6.1.9
moviepy>=1.0.3
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
instagrapi>=2.0.0
